add Quantize, various performance enhancements
Fixes: #44

Former-commit-id: 8178cfabffdaf937b079e41373394006f8bd8115 [formerly f94e71df39c437409bd7653a10929902aa0cf89b]
Former-commit-id: 41c893d95cd479ee144abf122d41f7319119efd3

+++ decimal_suite_test.go
@@ -152,7 +152,7 @@ func testCase(fname string, i int, c suite.Case, mode OperatingMode, t *testing.
 	}
 
 	want := dataToBig(c.Output, z.Context)
-	prec := int32(want.Precision())
+	prec := want.Precision()
 	if prec != 0 {
 		z.Round(prec)
 	}
@@ -228,7 +228,7 @@ func dataToBig(s suite.Data, ctx Context) *Big {
 	}
 }
 
-func shellOut(x, y *Big, op suite.Op, prec int32) *Big {
+func shellOut(x, y *Big, op suite.Op, prec int) *Big {
 	var opstr string
 	switch op {
 	case suite.Add:
